requirements.txt
rapidocr-onnxruntime
numpy
Pillow
openpyxl
pypdf
pdf2image
pypdfium2
